Remove need for SQLDelight primitive adapters
(cherry picked from commit cd91ea9b7723eec378bbbefd916d3da6407d685e)

# Conflicts:
#	app/build.gradle.kts
#	app/src/main/java/eu/kanade/tachiyomi/AppModule.kt
#	app/src/main/java/eu/kanade/tachiyomi/data/backup/BackupManager.kt
#	data/src/main/java/tachiyomi/data/manga/MangaRepositoryImpl.kt
#	domain/src/main/java/tachiyomi/domain/history/model/HistoryWithRelations.kt

+++ app/src/main/java/eu/kanade/tachiyomi/ui/browse/migration/advanced/process/MigratingManga.kt
@@ -30,11 +30,11 @@ class MigratingManga(
     }
 
     data class ChapterInfo(
-        val latestChapter: Float?,
+        val latestChapter: Double?,
         val chapterCount: Int,
     ) {
         fun getFormattedLatestChapter(context: Context): String {
-            return if (latestChapter != null && latestChapter > 0f) {
+            return if (latestChapter != null && latestChapter > 0.0) {
                 context.getString(
                     R.string.latest_,
                     DecimalFormat("#.#").format(latestChapter),
