+++ data/src/main/java/tachiyomi/data/manga/SearchTag.kt
@@ -2,13 +2,13 @@ package tachiyomi.data.manga
 
 import exh.metadata.sql.models.SearchTag
 
-val searchTagMapper: (Long, Long, String?, String, Int) -> SearchTag =
+val searchTagMapper: (Long, Long, String?, String, Long) -> SearchTag =
     { id, mangaId, namespace, name, type ->
         SearchTag(
             id = id,
             mangaId = mangaId,
             namespace = namespace,
             name = name,
-            type = type,
+            type = type.toInt(),
         )
     }
