+++ data/src/main/java/tachiyomi/data/manga/SearchTitle.kt
@@ -2,12 +2,12 @@ package tachiyomi.data.manga
 
 import exh.metadata.sql.models.SearchTitle
 
-val searchTitleMapper: (Long, Long, String, Int) -> SearchTitle =
+val searchTitleMapper: (Long, Long, String, Long) -> SearchTitle =
     { id, mangaId, title, type ->
         SearchTitle(
             id = id,
             mangaId = mangaId,
             title = title,
-            type = type,
+            type = type.toInt(),
         )
     }
